requirements.txt
pytest==7.3.2
selenium==4.10.0
faker==18.11.2
allure-pytest==2.13.2
requests==2.31.0
urllib3==2.0.6